Moved NMEA 0183 file
